Ok I finally have exactly the behavior I want...
It should not have taken me that long, if I had just focused on the
right details in Nuklear rather than getting hung up on the crazy
complicated parts that don't matter.  As well as noticing some issues
that were throwing me off, like the unused horizontal scrollbar taking
up some vertical space and the fact that it is not a truely free
scroll, but a stepwise scroll so the top element is always
perfectly aligned in the scroll range [list_index*row_height,
(list_index+1)*row_height).  Only the last element can be partially
displayed and given the stepwise nature that means the list has
extra space after the last element because that's the only way
to display the last element completely.